# [feature-wip] (memory tracker) (step5) Fix track bthread, fix track vectorized query