… be/src/runtime/thread_mem_tracker_mgr.cpp
… be/src/runtime/thread_mem_tracker_mgr.h
… be/src/service/brpc_conflict.h
… be/src/service/internal_service.cpp
… be/src/util/bit_util.h
… be/src/util/doris_metrics.cpp
… be/src/util/doris_metrics.h
… be/src/util/file_utils.cpp
… be/src/vec/common/allocator.h
… be/src/vec/exec/vexchange_node.cpp
… fe/fe-core/src/main/java/org/apache/doris/qe/SessionVariable.java